requirement.txt
aliyun-python-sdk-core==2.13.29
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.13.0
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
cos-python-sdk-v5==1.9.0
crcmod==1.7
cryptography==3.2.1
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django==2.2.12
django-crispy-forms==1.9.1
django-debug-toolbar==3.1.1
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.1.0
django-reversion==3.0.7
django-shortuuidfield==0.1.3
django-suit==0.2.28
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
drf-writable-nested==0.6.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.8
MarkupPy==1.14
mysql==0.0.2
mysqlclient==1.4.6
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
oss2==2.13.1
Pillow==7.1.2
public==2019.4.13
pycparser==2.20
pycryptodome==3.9.9
PyMySQL==0.9.3
python-memcached==1.59
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
requests==2.23.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
vine==1.3.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
